requirements.txt
pandas
openai
psycopg2
python-dotenv
instructor
anthropic
pgvector
numpy
tqdm
pandas